Change assembly location from Entry to Calling
Change required when directly calling the converts on the main process.
Making this change on the base Converter should work for both pipe system as well as new intended system, as the calling assembly for the base converter always will be the concrete converter for a particular version, hence the folder that the for example BHoMUpgrader72.exe lives in.