Turn off console logging
Turns off the console logging by default. More critical for console apps as the versioning is now called directly, leading to an bombardment of versioning logging drastically slowing the upgrader down.
When ProcessingLoop is called, the logging is turned on, as this is only done when the versioning is run as a standalone .exe with its own log, and not affecting the main log of any application.